+++ chirp_macros/src/generate.rs
@@ -11,7 +11,7 @@ enum FnConfig {
 const METHOD_ATTR_DESCR: &str = "\
 - `parse_dsl(ignore)`: Do not add this method to the parse_dsl_impl implementation
 
-There is currently no other accepted parse_dsl_impl method attribute config options.\n";
+There is currently no other accepted parse_dsl method attribute config options.\n";
 
 #[allow(clippy::trivially_copy_pass_by_ref)] // false positive. Type necessary to avoid eta-expension
 fn is_parse_dsl_attr(attr: &&syn::Attribute) -> bool {
@@ -20,11 +20,6 @@ fn is_parse_dsl_attr(attr: &&syn::Attribute) -> bool {
 impl FnConfig {
     #[allow(clippy::needless_pass_by_value)] // false positive. Type necessary for calling it
     fn parse(&mut self, meta: ParseNestedMeta) -> syn::Result<()> {
-        if *self != Self::default() {
-            let msg = "More than one `parse_dsl` meta attribute was declared \
-                for this method, can't know which one to chose! Only use one.";
-            return Err(meta.error(msg));
-        }
         match () {
             () if meta.path.is_ident("ignore") => {
                 *self = Self::Ignore;
@@ -50,10 +45,30 @@ impl FnConfig {
     }
 }
 
+// Similar to [`syn::Path::is_ident`] but checks for [`syn::Type::Path`] and
+// may work on generic types where `is_ident` doesn't.
+fn is_type(ty: &syn::Type, ident: &str) -> bool {
+    if let syn::Type::Path(ty) = ty {
+        ty.path.segments.last().is_some_and(|l| l.ident == ident)
+    } else {
+        false
+    }
+}
+struct TypeParser {
+    typ: syn::Ident,
+    parser: syn::Expr,
+}
+impl TypeParser {
+    fn is_type(&self, ty: &syn::TypePath) -> bool {
+        ty.path.is_ident(&self.typ)
+    }
+}
+
 pub(crate) struct ImplConfig {
     chirp_crate: syn::Path,
     delegate: Option<syn::Ident>,
     set_params: Option<syn::Generics>,
+    type_parsers: Vec<TypeParser>,
 }
 impl Default for ImplConfig {
     fn default() -> Self {
@@ -61,6 +76,7 @@ impl Default for ImplConfig {
             chirp_crate: syn::parse_quote!(::cuicui_chirp),
             delegate: None,
             set_params: None,
+            type_parsers: Vec::new(),
         }
     }
 }
@@ -73,6 +89,30 @@ const CONFIG_ATTR_DESCR: &str = "\
   in this `impl` block. This should be the field you mark with `#[deref_mut]`
 - `set_params <D: ParseDsl>`: Instead of re-using the `impl` block's generics \
   in the `impl<XXX> ParseDsl for Type` use the expression within parenthesis.
+- `type_parsers(<arg_type1> = <parser1>, <arg_type2> = <parser2>, …)`: \
+  For arguments of type `arg_type1`, use `parser1` a function of the following type:
+
+    fn parse(
+        registry: &TypeRegistry,
+        ctx: Option<&LoadContext>,
+        input: &'a str,
+    ) -> Result<ArgumentType, anyhow::Error>;
+
+To parse the argument. The default are as follow:
+
+- For `Handle<T>` and `&Handle<T>` arguments, `to_handle` is used.
+- For `&str` arguments, `identity` is used.
+- For any other type, `from_reflect` is used. It requires however that the \
+  argument type be `Reflect` and `FromReflect`.
+
+There are other options available:
+- `from_str`: it only requires the argument type to be `FromStr`
+- `<parser>` may accept arbitrary expressions, you can use your own parser as \
+  long as it has the type signature mentioned earlier. You can even define the \
+  parser as a closure inline.
+
+Currently, the type must be an identifier, so you can't handle (yet) generic types \
+and references this way.
 
 There is currently no other accepted parse_dsl_impl attribute config options.\n";
 
@@ -91,12 +131,22 @@ impl<'a> ImplConfig {
             () if meta.path.is_ident("set_params") => {
                 self.set_params = Some(meta.input.parse()?);
             }
+            () if meta.path.is_ident("type_parsers") => {
+                meta.parse_nested_meta(|meta| {
+                    let Some(ident) = meta.path.get_ident() else {
+                        return Err(meta.error("type_parsers type must be an identifier"));
+                    };
+                    let value = meta.value()?;
+                    self.type_parsers
+                        .push(TypeParser { typ: ident.clone(), parser: value.parse()? });
+                    Ok(())
+                })?;
+            }
             () => {
                 let path = &meta.path;
                 let ident = quote!(#path);
                 let msg = format!(
-                    "Unrecognized parse_dsl_impl meta attribute: \
-                    {ident}\n{CONFIG_ATTR_DESCR}"
+                    "Unrecognized parse_dsl_impl meta attribute: {ident}\n{CONFIG_ATTR_DESCR}"
                 );
                 return Err(meta.error(msg));
             }
@@ -126,10 +176,10 @@ pub(crate) fn parse_dsl_impl(config: &ImplConfig, block: &mut syn::ItemImpl) -> 
     let this_crate = &config.chirp_crate;
 
     let funs = block.items.iter().filter_map(dsl_function);
-    let funs = funs.map(method_branch);
+    let funs = funs.map(|f| method_branch(f, &config.type_parsers));
     let catchall = config.delegate.as_ref().map_or_else(
         || quote!(Err(DslParseError::<Self>::new(name))),
-        |ident| quote!(self.#ident.method(MethodCtx { name, args })),
+        |ident| quote!(self.#ident.method(MethodCtx { name, args, ctx, registry })),
     );
     let parse_dsl_block = quote! {
         #[automatically_derived]
@@ -140,7 +190,9 @@ pub(crate) fn parse_dsl_impl(config: &ImplConfig, block: &mut syn::ItemImpl) -> 
                 data: #this_crate::parse::MethodCtx,
             ) -> Result<(), #this_crate::anyhow::Error> {
                 use #this_crate::parse::{quick, MethodCtx, DslParseError};
-                let MethodCtx { name, args } = data;
+                use #this_crate::wrapparg::{from_str, from_reflect, to_handle, identity};
+
+                let MethodCtx { name, args, ctx, registry } = data;
                 match name.as_ref() {
                     #(#funs)*
                     _name => { #catchall }
@@ -156,7 +208,7 @@ pub(crate) fn parse_dsl_impl(config: &ImplConfig, block: &mut syn::ItemImpl) -> 
     quote!(#block #parse_dsl_block)
 }
 // Note: assumes cuicui_chirp::parse::quick is in scope and used correctly
-fn method_branch(fun: &syn::ImplItemFn) -> TokenStream {
+fn method_branch(fun: &syn::ImplItemFn, parsers: &[TypeParser]) -> TokenStream {
     match FnConfig::parse_list(&fun.attrs) {
         Ok(FnConfig::Ignore) => return TokenStream::new(),
         Ok(FnConfig::Method) => {}
@@ -174,12 +226,33 @@ fn method_branch(fun: &syn::ImplItemFn) -> TokenStream {
     let index = syn::Index::from;
     let quote_arg = |i: syn::Index| if arg_count == 1 { quote!(args) } else { quote!(args.#i) };
     let fun_args = (0..arg_count).map(index).map(quote_arg);
+    let arg_parsers = fun.sig.inputs.iter().skip(1);
+    let arg_parsers = arg_parsers.map(|a| argument_parser(a, parsers));
     let ident = &fun.sig.ident;
     quote_spanned! { fun.sig.inputs.span() =>
         stringify!(#ident) => {
             let args = quick::#arg_n(args.as_ref())?;
-            self.#ident(#(#fun_args),*);
+            self.#ident(#(#arg_parsers(registry, ctx, #fun_args)?),*);
             Ok(())
         }
+    }
+}
+fn argument_parser(argument: &syn::FnArg, parsers: &[TypeParser]) -> TokenStream {
+    use syn::Type::{Path, Reference as Ref};
+    use syn::TypeReference as TRef;
+
+    match argument {
+        syn::FnArg::Receiver(_) => unreachable!(),
+        syn::FnArg::Typed(syn::PatType { ty, .. }) => match ty.as_ref() {
+            Path(ty) if parsers.iter().any(|prs| prs.is_type(ty)) => {
+                let find = |prs| TypeParser::is_type(prs, ty).then_some(&prs.parser);
+                let parser = parsers.iter().find_map(find).unwrap();
+                quote!(#parser)
+            }
+            Path(ty) if ty.path.is_ident("Handle") => quote!(to_handle),
+            Ref(TRef { elem, .. }) if is_type(elem, "Handle") => quote!(&to_handle),
+            Ref(TRef { elem, .. }) if is_type(elem, "str") => quote!(identity),
+            _ => quote!(from_reflect),
+        },
     }
 }